requirements.txt
# Bindings for the hermes protocol 
hermes-python>=0.3.2

requests
inflect
python-Levenshtein
fuzzywuzzy
paho-mqtt
ourgroceries



